add implementation StepExecutionClickXpath
