+++ src/main/java/executor/Application.java
@@ -1,0 +1,13 @@
+package executor;
+
+/**
+ * Hello world!
+ *
+ */
+public class Application
+{
+    public static void main( String[] args )
+    {
+
+    }
+}
