--- src/main/java/executor/service/Application.java
@@ -1,15 +1,0 @@
-package executor.service;
-
-import executor.service.model.Step;
-
-/**
- * Hello world!
- *
- */
-public class Application
-{
-    public static void main( String[] args )
-    {
-
-    }
-}
