--- src/main/java/executor/service/StepExecution/WebDriverChrome.java
@@ -1,18 +1,0 @@
-package executor.service.StepExecution;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-
-public class WebDriverChrome implements WebDriverInitializer{
-
-    Property property = new Property();
-
-    @Override
-    public WebDriver create() {
-        property.connect();
-        ChromeOptions options = new ChromeOptions();
-
-        return new ChromeDriver();
-    }
-}
