--- src/main/java/executor/service/StepExecution/WebDriverInitializer.java
@@ -1,7 +1,0 @@
-package executor.service.StepExecution;
-
-import org.openqa.selenium.WebDriver;
-
-public interface WebDriverInitializer {
-        WebDriver create();
-}
