+++ src/main/java/executor/service/WebDriverInitializerService.java
@@ -1,0 +1,7 @@
+package executor.service;
+
+import org.openqa.selenium.WebDriver;
+
+public interface WebDriverInitializerService {
+        WebDriver create();
+}
