+++ src/test/java/executor/ApplicationTest.java
@@ -1,0 +1,24 @@
+package executor;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+/**
+ * Unit test for simple App.
+ */
+public class ApplicationTest
+{
+    /**
+     * Rigorous Test :-)
+     */
+
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue(true );
+    }
+
+
+}
