--- src/test/java/executor/service/ApplicationTest.java
@@ -1,26 +1,0 @@
-package executor.service;
-
-import executor.service.model.ThreadPoolConfig;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-/**
- * Unit test for simple App.
- */
-public class ApplicationTest
-{
-    /**
-     * Rigorous Test :-)
-     */
-
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue(true );
-    }
-
-
-}
