--- src/test/java/executor/service/model/ProxyCredentialsTest.java
@@ -1,20 +1,0 @@
-package executor.service.model;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class ProxyCredentialsTest {
-
-@Test
-    public void login(){
-        ProxyCredentials login = new ProxyCredentials();
-        ProxyCredentials actual = new ProxyCredentials("Vlad","1111");
-        login.setUsername("Vlad");
-        login.setPassword("1111");
-
-
-        assertEquals(actual,login);
-    }
-
-}
